rating-system.com rating and review plugin
Added partial with a rating-system.com widget configured for testing.
Commented out the render of the partial.
It will be too clunky to harvest the average rating value to put in our
search index.  We can more easily do it with their API (I think), but
that costs too much money at this time.