Refactored provider registration to facilitate client registration.
Reverted to using the default Devise registration controller.
Now use a hidden field in the registration form to determine if the new
user should be a provider (expert).
The user model will add the expert role if the is_provider attribute is
set to a "present" value and the user is not already a client.