Implemented facility for a provider to claim their profile.
A profile editor can email an invitation to a provider requesting them
to claim a previously created profile that has not yet been claimed.
The email contains a link that will attach the profile to the account
of the logged-in user, provided that user currently has no profile.  If
the email recipient is not previously registered and clicks the link,
then they can register and will be redirected to the URL that will
attach the profile to their new account.
Fixed sunspot config bug in profile model: a nil last_name attribute
should be converted to empty string when defining scoping for
last_name.  Do the same when sorting by last_name in the profile index
view.