Geocode locations.
When a location is saved, geocode the address (get latitude and
longitude).
Until we go international, do not include country in display address or
geocodable address.
If geocodable address is empty, explicitly set coordinates to nil.
Cleaned up display address for case when we only have a postal code to
show.