Copy profile primary phone to location phone.
For any profiles that have a primary phone in their profile and no or
one location, copy the phone number into the one location (creating the
location if needed).
Re-expose the profile primary and secondary phone attributes to be used
for manually copying phone numbers in profiles that have multiple
locations.