Sign-in or sign-up from a profile's rating area returns to the profile.
If a new visitor on a profile page wants to rate that provider, they
are prompted to sign in or sign up.  If they do so, they will be
redirected back to the provider's profile after signing-in or signing
up.
Allow for nil or new user in Profile#rating_by.