Cucumber features for provider profile access with new design.
Align features and steps to reflect a provider viewing and editing
their own profile using the new design.  Many fixed, but more to do.
Now that viewing and editing a profile requires it to be persisted,
augment profile claim process to allow a provider to replace their
existing profile with a claimed profile.
When redirecting, use a full URL, not a path.
Displays of attribute values on the profile edit page now have the
attribute_display CSS class.