Made navigation better for providers and other stuff.
Moved "My profile" link to top of the page and also on the home page.
Don't need to pass in profile explicitly to generate links to view or
edit my profile.
Don't call preserve directly on user-input data unless we know it's
safe.
Prominently prompt the provider, when viewing their own profile, to
edit their profile.