Browse by Category on home page.  Categories and services admin.
Implemented "Browse by Category" on the home page. Placing of
categories and services on the home page is authorable.
On a category's edit page, we can edit the display order and names of
its services.
On a service edit page, we can edit the names of its specialties.
Removed from the profile search index: consultation modes and profile
attributes that we are no longer using.
Added hours to the profile search index.