Sign-up and sign-in pages on new design.  Bug fix for sorting profile locations.
Converted sign-up and sign-in pages to new design.
The layout for the new design is now the default.
Sort profile locations by ID.  It's the same result as using created_at
and is more efficient.
On profile edit page when displaying the locations formlet, assume that
one or more locations have not been saved yet (due to a error on
creation).  This means that we have to check for nil when sorting
locations by ID here.  Also, we can't use a query scope to order
locations here.