Ensure search results pop-up top is not above viewport.
Deal with an edge case for placing the pop-ups on the search results
page.  If we put the pop-up above the entry, make sure it doesn't go
off the top of the viewport.  If it does, place its top just below the
viewport top.  In this case, the pointer might now be over the pop-up
and if it is, use the mouseleave event of the pop-up to hide it.