Profile editor can fully manage profiles.
Profile editor can now do anything to profiles that an administrator
can.
Test profile permissions using CanCan instead of checking the user
role. The exception is for publishing and setting admin attributes in
which case we also check if the user is a profile editor (we need to
make sure a provider cannot publish their own profile).