Bug fix to admin editing of categories and services.
Use sort_by when doing case-insensitive sort by the name of an object.
Hoping this is more efficient.
When editing a category or service, include already assigned services
or specialties, respectively, in the list that can be assigned, even if
they are no longer predefined.  This removes the need to do things in
the correct order when retiring a service or specialty.