Character count on custom services and specialties input fields.
Added character counts to the custom services and custom specialties
fields on the profile edit page.
Modified simplyCountable jquery plug-in.  Added ability to modify the
options on a per element basis, e.g., for all rendered custom services
inputs after a validation error.  Added option to append to the list of
events that trigger a count update.