Browse-by-category links search by service name as an exact phrase.
When searching with service_id param, uses Profile.search_by_service
method.  This does a search on the service name as an exact phrase
while boosting matches in the services text field.  This results in
search results with profiles that have that service assigned first in
the list, then profiles with a match on the name as a phrase later in
the list.
Refactored the profile search code to do the above.
Profile.fuzzy_search behavior is unchanged.  Added the
Profile.configurable_search method which is called by the two previous
search methods.