Locations formlet fix on profile edit page.
In the locations formlet of the profile edit page, if the stored
profile has no locations, then we create a dummy location for display
purposes.  The controller's build of an initial location doesn't help
here because we are looking at the stored locations to display the
location attributes on the edit page (using the order_by_id scope).
When displaying the locations, we don't need the 'try' method to access
location attributes, because we are ensured of having a location object.