Added ratings to reviews in a one-to-one association.
Each review now has a one-to-one association with a rating.  Ratings
for a profile are now accessed through the reviews.
Rating scores are now integers.  The average rating score for a profile
is still a float.
Track changes to reviews.
Removed ability to access the deprecated profile rating mechanism.
Profile page now has a link to the Google review form.  It prefills the
provider name, as well as user info if available.
Fixed glitch in logo display on layout header.