Allow longer reviews and strip whitespace from reviewer email.
We need 4000 characters for the review body.
Strip leading and trailing whitespace from input reviewer email,
otherwise you can get a spurious uniqueness error on  the username if
the input email has whitespace and the reviewer is already in the
database.
Stripping whitespace from the input reviewer email revealed strange
uniqueness errors on the reviewer attributes when creating reviews with
FactoryGirl.  Did a work around.