See all categories page.
Created new page that shows all available categories and services.
Modified home page browse-by-category to allow admin to specify a
subset of all categories and services to display.
Admin can specify which columns for categories to appear in on both
pages.
Simplified "predefined" scope for categories, services, and
specialties: makes it easier to specify sorting order with multiple
scopes.