Users can rate a provider. New layout for ratings and review link.
Reintroduced stars on the profile page for rating a provider.
Redid the layout for rating stars and the review link on the profile
view and edit pages, as well as search results.
Removed contact icons from search results.
Only show review forms on the admin edit page if the current user is
allowed to create and edit reviews.