Rating on the profile page sends you to a review form.
After rating a provider from their profile page we now redirect to the
review form.  If the user is not signed in, then they go to the sign-up
page.  After sign-up or sign-in, they will be redirected to the profile
rate URL, which will subsequently redirect them to the review form.
Modified profiles_controller#rate to handle non-AJAX calls.  Also fixed
the AJAX handling to do proper JSON in both the client-side Javascript
and the controller.
Fixed rate partial to display proper prompts on both the profile and
review pages.