Made ActiveRecord counter cache more robust.
Fixed the counter cache for associations to update when the record that
"belongs to" is updated.  This allows for creating a record before
associating it with the parent.
The counter cache value is retrieved from the actual number of
associated records, rather than relying on increments and decrements
which can be subject to data skew.
Added counter cache to the Location model (I had missed it before).
The debugger for the development environment.