Switch to all key-based caching.
Switched to key-based caching.
Use cache_digests to detect changes in the template portions that are
encompassed by fragment caching.
Cache sweepers and action caching are deprecated with Rails 4, so let's
not use them.
Created category_lists. They can aid in key-based caching for category
lists on the home page and see-all categories page.
Dynamically ensure proper category membership in the "home" and "all"
category lists.
Touch the appropriate category_lists when modifications are made to
member categories and services to aid in key-based caching.