Work-around for optional Aviary JS file loading.
Due to an enclosing fragment cache, we can't use content_for to add a
Javascript include tag in the photo-upload partial.  Instead use a
boolean flag outside of the cache fragment to instruct the layout to
add the include tag.
Added a couple of scenarios for clicking on the photo on the profile
edit page to bring up the photo-upload pop-up.