Fix bug in services formlet caused by caching on profile edit page.
Caching of the profile edit page introduced a bug in the services
formlet.  When categories, services, or specialties are modified, the
changes need to appear on cached profile edit pages.  Fixed by loading
the needed info in an AJAX call.  Also, needed to isolate the services
formlet from fragment caching.