Supply missing placeholder image for the profile photo.
Because we are serving our static image assets from our server, we need
all references to them to exist (otherwise we get a render error and an
error page on production).
Created a large placeholder image.
Need the medium placeholder to match the pop-up size on the search
results page.