Get scenarios working that we are now running with Javascript.
Upgraded to latest version of email_spec helpers.
Add a delay to the email_spec helpers that we are using.  Apparently
for steps that run with Javascript, we need to wait a bit before
attempting to access the email.
Removed a claim-profile scenario that is now impossible.
When specifying a link to visit, use the path, not the full URL.