In private-site mode, don't send confirmation email upon registration.
Added configuration parameter for running the site in private mode.
If we're running in private-site mode, do not automatically send a
confirmation email upon registration.  Instead, an admin user will
trigger it manually if desired.  This allows visitors to request access
by registering and will be able to later sign-in upon receipt of a
confirmation email.
Added user attributes for tracking delivery of the confirmation email
by an admin user.