Cucumber tweaks to deal with new pop-over closing.
Consolidate "click on" step definitions in the site-wide file.
The new javascript for closing profile edit formlets is apparently not
handled well by capybara-webkit.  When trying to use a second formlet
in one step, the click of the second formlet is not recognized.  So
let's stick with using only one formlet at a time.