More search modifications.
Various modifications in an attempt to eliminate irrelevant results:
* Restrict full-text profile search to name, company, headline,
languages, addresses, categories, services, specialties, specialty
search terms, and profile search terms.
* For non-phrase full-text searches, tweaked the "mm" parameter to
require more of the clauses in the query.
* Removed the PhoneticFilter from the analyzer list.  It was matching
way too many irrelevant words. That's too bad because it would be
useful for hard-to-spell words.

Removed customization of default query parameters.  We can't set this
on production and it was causing slightly different results between
development and production.

Added admin display of categories, services, and specialty search terms
to the profile show page.  This can be used to understand why a
particular profile was returned by a search.