Provider edits specialties only...
The provider no longer edits the categories or services associated with
their profile.  Leaving that to admin.
Provider adds, creates, or removes specialties using a pop-over with
input text fields (no more check boxes).  No more dependency on which
categories or services are associated with the profile.
Work around to make the rails3-jquery-autocomplete code not show a
selection menu if there are no matches to the input text.