Tweaks to search results redesign.
"More details" link needs separation from a long provider name.
Use cursor pointer when hovering over the provider photo.
Search results title is not safe to render raw because it includes the
tainted query string.
Account drop-down menu should link to "nice" profile show and edit URLs.
Both input element IDs for the search form need to match what's used by
Cucumber steps and the location input ID needs to match what we are
using for Google Places autocompletion.  Also, these IDs should use
unique names because they are in the layout and generic names are
likely to conflict with elements in the content area.
If we can get away with it, let's not mix interpolated strings with
HTML in en.yml.  That way we don't have to worry about escaping the
interpolated values ourselves for security reasons.