Implemented email opt-out via ContactBlockers.
Profile invitation email now has an email opt-out link.  Recipient can
opt-out of future emails as long as they click the link in the email
they received.  Makes it harder for nefarious people to opt out other
people.
Admin cannot send an invitation for a profile if a recipient of a
previous invitation for that profile opted out.