Cucumber scenarios conform to new design.
Fixed the Cucumber scenarios and step definitions to conform to the new
design.
The custom Modernizr we got from Andy was missing AJAX and file reader
detection and thus was incorrectly telling us not to display the file
attachment input.  I added the file-api and network-xhr2 capabilities
to the custom Modernizr.