Fix to user edit by admin.
users_controller is for editing other users by admin, not for editing
the current_user. users/registrations_controller is for editing the
current_user.
The exception is that this controller is used to claim a profile--a
legacy of a previous design, I think.
The view_profile, update_profile, edit_profile actions are obsolete
code, I think.