Work around widget JS error on profile edit tab.
Given that we don't know what will be in the widget code, let's not try
to display the widget in both the profile show and edit tabs, i.e.,
rendering the widget code twice on the same page might give
unpredictable results.  The work-around is to only display the widget
in the show tab.
https://www.pivotaltracker.com/story/show/79159586