More work for provider charges to customers.
Removed $1.00 charge when creating a Stripe customer and card.  It's
not needed.
Added ability for a customer to change the card used by a particular
provider.  The new card will also be added to the customer's list of
available cards for use with other providers.
Added ability for customer to increment the authorized amount for a
particular provider.
Renamed customer_file attribute from authorization_amount to
authorized_amount.