Transactional emails less spammy for gmail.
Only show logo image for the welcome and invitation to claim emails --
they have lots of text.  Emails with a low text to image ratio are more
likely to be marked as spam.
Go back to sending from the support address.
Use a subject line in the confirmation emails that does not make gmail
mark it as spam.