Require username when reviewing.
When any user that doesn't already have a username is writing a review,
show them a username field on the review form and require it to be
filled in.
When admin is creating or editing a review, require a username for the
reviewer, even if the reviewer is a provider.