Display a user's account with users#show.
We've been viewing a user's account with users#edit in the hopes that
editing would be implemented.  Accepted reality and implemented the
user account page with the show action.  That's the correct way to do
it anyway.
Modified all references to the user account page to use the show action.