Charge refund tweaks.
Record the date and time of the last refund.
In the refund notification email, show the amount of the refund that
was just applied, not the total refund amount (if there were multiple
refunds on this charge).
To the provider, show the original amount charged, and if there was a
refund, the total refund amount, last refund date, and the amount
collected.
To the client, show the last refund date.
Tagged all Cucumber scenarios and Rspec examples that are related to
payments.