After-sign-in path should not persist.
Use flash to set the after_sign_in_path_override.  Setting it in the
session caused it to interfere with the stored location value, e.g.,
when later attempting to log-in or sign-up to review a provider.  See
https://www.pivotaltracker.com/story/show/86391590.
Don't set the after_confirmation_url session value when visiting the
newsletter subscription page.  That will also interfere with the stored
location.  It doesn't need to be set when subscribing to the newsletter
as far as I can tell.
Use the anchor parameter when adding a fragment ID to a path or URL.