Edit subscriptions without entering a password.
Use the users_controller to update newsletter subscriptions.  This
allows us to bypass Devise and modify subscriptions without entering a
password on the submission form.
Added subscriptions to the account drop-down menu.
Wording tweaks on the subscription management form.
Modified the user model to allow access to a subset of attributes
without entering a password on the submission form.
Cleaned up the users_controllers and user Rspec examples.