Added Stripe Connect setup to profile edit tab.
Include Stripe Connect setup on the profile edit tab if the profile
belongs to the current user and the user is not already connected to
Stripe.
Share a partial for Stripe Connect setup between the settings page and
profile edit tab.
When the Stripe Connect button is clicked, use the current window to
display the setup form rather than using a pop-up window.  The code
will support a pop-up if we want to revert to that.