Sign-up only has email, password & subscriptions.
Remove the postal code, username, and password confirmation inputs from
all sign-up forms.
After successfully submitting a change on the Settings page, redirect
back to the same page (instead of the home page or profile page).
On the Settings page, inform the user if an email change confirmation
is pending.
Always show the Settings and Subscriptions links in the Account menu at
the top of each page.
If Stripe Connect form is cancelled or gets an error, have the
confirmation page button return to the place the provider clicked the
Stripe Connect button.  If successful, the button will always go to the
profile edit tab.