Google Analytics tracking config and user ID
Configure the Google Analytics tracking ID from the environment.  Use
auto-configuration of the tracking cookie domain. This allows us to set
up test tracking for the development and staging environments.
The user ID we were using was not stable.  Use a hex digest instead
which should always be the same for each user.