Newsletter sub does not require confirmation.
Email confirmation is no longer required to subscribe the user to a
mailing list.  As soon as the user submits the registration form, they
will be subscribed.  However, they still need to confirm their email to
complete account registration.
Modified the confirmation pages and email for newsletter sign-up on the
main site and blog to reflect the above.
Show subscriptions in the registration notification emails and user
account page in the admin interface.