Newsletter sign-up config tweaks and delay fix.
Set up the NewsletterSubscriptionJob class in its own file.  If it's
configured in the newsletters controller, when delayed_job instantiates
the controller it will likely get an error trying to set the
NewsletterSubscriptionJob constant again.  Also, need the method that
does the work to be a public class method
(NewslettersController.subscribe_to_mailing_lists).
Allow the MailChimp welcome email and background processing of the
mailing list subscriptions to be configured from the environment.
Makes it easier to test during development.
Don't send the MailChimp welcome email in the test environment.
The nlsub parameter set in the routing configuration doesn't get added
as a query parameter, so set it at run time.