Caching fix on newsletter archive page.
If we have cached content that depends on the login state, we need to
put that in the cache key.
However, let's always show the link back to the newsletter sign-up
page, even if logged in.  We may want to make the sign-up page update
the user record if they are logged in.  And let's put the login state
in the cache key anyway in case we add login-dependent content later.