Upgraded Capybara Webkit to 1.6.0.
The xpath expression we were using to find the page title in a step
definition was not working with the new Webkit.  Fortunately, we can
access the page title simply with the "title" method.
Also removed the tagline from page titles.  It might have been making
page titles too similar to each other and Google doesn't like that.
Unique page title for the newsletter sign-up page.