Fix truncated specialties in search results
If a provider has a long list of specialties, we truncate it in the
search results display.  The Rails 4 truncate method was escaping the
span tags in which we embed each specialty.  Fixed by preventing the
truncate method from escaping and using the sanitize method with an
exception for span tags.