Better handling of unsubscribe errors
If the unsubscribe link given to the recipient is corrupted, send them
to the blank opt-out form (instead of showing a dead-end error page).
On the unsubscribe error page (reached from a corrupted form
submission), give them the option to try the blank opt-out form.
Improvements to the email deliveries admin interface.