Update user "leid" value when email address changed via MailChimp
If a subscriber with a user account (or an admin) changes the
subscription’s email address via MailChimp, we do not want to change
the email address of the user account.  However, we still need a way to
tie the user account to the subscription on MailChimp (which now has a
new email address). Do this by updating the user account’s “leid” value
with the new unique_email_id.  That way we can use that value to find
the subscription on MailChimp if the user account modifies the
subscription or vice versa.