Allow for campaign still sending during MailChimp web-hook notification
A MailChimp web-hook notification for a “sent” campaign can arrive
while the campaign is still in the process of sending. Allow for this.
Mark stage-based campaigns as sent even if they are still in the
process of sending. It’s not worth creating a robust system for waiting
to see if the campaign finishes sending.