Work-around to ensure that payment profile announcements are displayed
For some reason unknown to me, ActiveRecord does not always query for
the PaymentProfileAnnouncement sub-type when looking up a profile’s
ProfileAnnouncements. If we first query for a profile’s
PaymentProfileAnnouncements via profile.payment_profile_announcements,
then it will query for that sub-type when looking up
profile.profile_announcements.
Also, when displaying search results, do not query for announcements if
a profile’s cache fragment is still valid.